# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sahara Triage Voice Agent API",
    "summary": "Conversational health-triage voice agent: Supabase-backed triage, scheduling, and clinician workflows.",
    "description": "Patient signup and voice triage (browser WebSocket or Africa's Talking phone channel), automatic doctor-slot routing, clinician queues with embedded triage reports, prescriptions, and consent-gated call workflows.\n\nAuth: Supabase session JWT as `Authorization: Bearer <token>` (use the Authorize button). The browser voice channel is a WebSocket and is documented in the README message catalog; phone callbacks accept `application/x-www-form-urlencoded` and return TwiML XML.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.sahara.example.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/auth/patient/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign up as a patient",
        "operationId": "patient_signup_api_auth_patient_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientSignup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientSignupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/doctor/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign up as a doctor (dummy license verify)",
        "operationId": "doctor_signup_api_auth_doctor_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorSignup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorSignupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/appointments": {
      "get": {
        "tags": [
          "appointments"
        ],
        "summary": "List my appointments",
        "description": "Patient-facing: their own appointment dates, most recent first.",
        "operationId": "list_my_appointments_api_appointments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/appointments/route": {
      "post": {
        "tags": [
          "appointments"
        ],
        "summary": "Auto-book earliest free doctor slot",
        "description": "Book the earliest free doctor slot for a triaged conversation.\n\nCalled by the web-voice client after TRIAGE_UPDATE; the phone channel\ncalls the same function internally in voice_webhook.py.",
        "operationId": "route_my_appointment_api_appointments_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteAppointmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteAppointmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppointmentListResponse": {
        "properties": {
          "appointments": {
            "items": {
              "$ref": "#/components/schemas/AppointmentOut"
            },
            "type": "array",
            "title": "Appointments"
          }
        },
        "type": "object",
        "required": [
          "appointments"
        ],
        "title": "AppointmentListResponse"
      },
      "AppointmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "tier": {
            "type": "string",
            "title": "Tier",
            "examples": [
              "AMBER"
            ]
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "doctor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "scheduled"
            ]
          },
          "notified_via": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notified Via"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "tier",
          "scheduled_at",
          "status"
        ],
        "title": "AppointmentOut"
      },
...
      "DoctorSignup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "license_number": {
            "type": "string",
            "title": "License Number"
          },
          "specialty": {
            "type": "string",
            "title": "Specialty"
          },
          "confirm_license": {
            "type": "boolean",
            "title": "Confirm License"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone_number",
          "email",
          "password",
          "license_number",
          "specialty",
          "confirm_license"
        ],
        "title": "DoctorSignup"
      },
      "DoctorSignupResponse": {
        "properties": {
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token",
            "description": "Supabase session JWT. Null until email confirmed."
          }
        },
        "type": "object",
        "required": [
          "doctor_id"
        ],
        "title": "DoctorSignupResponse"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Not a recognized patient session"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientSignup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone_number",
          "email",
          "password"
        ],
        "title": "PatientSignup"
      },
      "PatientSignupResponse": {
        "properties": {
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token",
            "description": "Supabase session JWT. Null when 'Confirm email' is enabled until the user confirms."
          }
        },
        "type": "object",
        "required": [
          "patient_id"
        ],
        "title": "PatientSignupResponse"
      },
...
      "RouteAppointmentRequest": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Conversation Id"
          }
        },
        "type": "object",
        "required": [
          "conversation_id"
        ],
        "title": "RouteAppointmentRequest"
      },
      "RouteAppointmentResponse": {
        "properties": {
          "appointment": {
            "$ref": "#/components/schemas/AppointmentOut"
          }
        },
        "type": "object",
        "required": [
          "appointment"
        ],
        "title": "RouteAppointmentResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Supabase session JWT",
        "scheme": "bearer"
      }
    }
  },
...
}